feat(deps): deprecate dynamic import <Dialog>
